+++ internal/httpapi/api_dialtesting.go
@@ -10,16 +10,12 @@ import (
 	"fmt"
 	"io/ioutil"
 	"net/http"
-	"os"
-	"path/filepath"
 	"strings"
 	"time"
 
 	dt "github.com/GuanceCloud/cliutils/dialtesting"
 	uhttp "github.com/GuanceCloud/cliutils/network/http"
-	"github.com/gin-gonic/gin"
 	"gitlab.jiagouyun.com/cloudcare-tools/datakit/internal/datakit"
-	"gitlab.jiagouyun.com/cloudcare-tools/datakit/internal/path"
 )
 
 type dialtestingDebugRequest struct {
@@ -120,89 +116,4 @@ func getAPIDebugDialtestingRequest(req *http.Request) (*dialtestingDebugRequest,
 	}
 
 	return &reqDebug, nil
-}
-
-func apiUploadSourcemap(c *gin.Context) {
-	context := getContext(c)
-
-	// check token valid
-	token := c.Query("token")
-	if len(token) == 0 {
-		context.fail(dcaError{Code: 401, ErrorCode: "auth.failed", ErrorMsg: "auth failed"})
-		return
-	}
-	localTokens := apiServer.dw.GetTokens()
-	if len(token) == 0 || len(localTokens) == 0 || (token != localTokens[0]) {
-		context.fail(dcaError{Code: 401, ErrorCode: "auth.failed", ErrorMsg: "auth failed"})
-		return
-	}
-
-	uploadSourcemap(c)
-}
-
-func uploadSourcemap(c *gin.Context) {
-	context := getContext(c)
-
-	var param rumQueryParam
-
-	if c.ShouldBindQuery(&param) != nil {
-		context.fail(dcaError{ErrorCode: "query.parse.error", ErrorMsg: "query string parse error"})
-		return
-	}
-
-	if param.ApplicationID == "" {
-		context.fail(dcaError{ErrorCode: "query.param.required", ErrorMsg: "app_id required"})
-		return
-	}
-
-	if param.Platform == "" {
-		param.Platform = SourceMapDirWeb
-	}
-
-	if param.Platform != SourceMapDirWeb && param.Platform != SourceMapDirMini &&
-		param.Platform != SourceMapDirAndroid && param.Platform != SourceMapDirIOS {
-		l.Errorf("platform [%s] not supported", param.Platform)
-		context.fail(dcaError{
-			ErrorCode: "param.invalid",
-			ErrorMsg:  fmt.Sprintf("platform [%s] not supported, please use web, miniapp, android or ios", param.Platform),
-		})
-		return
-	}
-
-	file, err := c.FormFile("file")
-	if err != nil {
-		l.Errorf("get file failed: %s", err.Error())
-		context.fail(dcaError{ErrorCode: "upload.file.required", ErrorMsg: "make sure the file was uploaded correctly"})
-		return
-	}
-
-	fileName := GetSourcemapZipFileName(param.ApplicationID, param.Env, param.Version)
-	rumDir := filepath.Join(GetRumSourcemapDir(), param.Platform)
-	if !path.IsDir(rumDir) {
-		if err := os.MkdirAll(rumDir, datakit.ConfPerm); err != nil {
-			context.fail(dcaError{
-				ErrorCode: "dir.create.failed",
-				ErrorMsg:  "rum dir created failed",
-			})
-			return
-		}
-	}
-	dst := filepath.Clean(filepath.Join(rumDir, fileName))
-
-	// check filename
-	if !strings.HasPrefix(dst, rumDir) {
-		context.fail(dcaError{
-			ErrorCode: "param.invalid",
-			ErrorMsg:  "invalid param, should not contain illegal char, such as  '../, /'",
-		})
-		return
-	}
-
-	if err := c.SaveUploadedFile(file, dst); err != nil {
-		l.Errorf("save upload file error: %s", err.Error())
-		context.fail(dcaError{ErrorCode: "upload.file.error", ErrorMsg: "upload failed"})
-		return
-	}
-	runSourcemapCallback(dst)
-	context.success(fmt.Sprintf("uploaded to [%s]!", dst))
 }